Moved functions to more relevant files
Moved changeLights to controls.py and moved loadDev, loadStates, and
initDev to utils.py

Testing: Tested with night.py and all appears to be working correctly
